It is more convinient for handlers to be on main export.
This allows for easier access right off the main export. In the future
these should become named exports to support spec complient `import`
objects.

See discussion
https://github.com/60frames/webpack-hot-server-middleware/pull/40#discussion_r138642046
