+++ src/index.js
@@ -5,7 +5,22 @@ const path = require('path');
 const requireFromString = require('require-from-string');
 const MultiCompiler = require('webpack/lib/MultiCompiler');
 const sourceMapSupport = require('source-map-support');
-const { createConnectHandler } = require('./handlers');
+
+const createConnectHandler = (error, serverRenderer) => (req, res, next) => {
+    debug(`Receive request ${req.url}`);
+    if (error) {
+        return next(error);
+    }
+    serverRenderer(req, res, next);
+};
+
+const createKoaHandler = (error, serverRenderer) => (ctx, next) => {
+    debug(`Receive request ${ctx.url}`);
+    if (error) {
+        ctx.throw(error);
+    }
+    return serverRenderer(ctx, next);
+};
 
 const DEFAULTS = {
     chunkName: 'main',
@@ -135,5 +150,7 @@ function webpackHotServerMiddleware(multiCompiler, options) {
         return options.createHandler(error, serverRenderer).apply(null, arguments);
     };
 }
+
+Object.assign(webpackHotServerMiddleware, { createConnectHandler, createKoaHandler });
 
 module.exports = webpackHotServerMiddleware;
